stacks: unify and cleanup landing checks system (Bug 1888188) (#388)
This commit begins the process of unifying and refactoring the
landing checks system, which blocks stacks and revisions from
landing and issues warnings about various issues in the landing
that are not blockers. The state of this revision gives us a
base to work on for extending the lints system to add Mercurial
commit hooks, clean up some unnecessary data fields, and improve
the user experience around presentation of landing lints.

The existing checks are split up into three categories. First are
stack-level blocker checks, which inspect the entire state of the
stack and block landings entirely. The second are revision-level
blocker checks which inspect the state of individual revisions and
block landing of each revision and it's successor nodes. Finally
are the revision-level warnings, which do not block landings but
will present information about sub-optimal landing conditions that
must be acknowledged before landing.

Each piece of state used by individual checks is added to a new
TransplantAssessmentState class. An instance of this class is
passed to each lint to keep the API consistent and discoverable.
Some fields on this class have been moved directly from existing
checks, and future refactors should reduce the number of fields on
this class.

The transplant assessment API and the calculate_landable_subgraphs
API are combined into a single API as part of this commit. Since
the previous setup had two different use cases, some state is not
required to run landing checks on non-transplant assessments. All
pieces of state required to check for a valid landing request are
encapsulated in LandingAssessmentState, which is an optional field
on TransplantAssessmentState. Checks which should only run on
landing assessments can look for the existance of a LandingAssessmentState
on the TransplantAssessmentState.

The calculate_landable_subgraphs API and the custom graph walking
code is replaced by the new APIs. When building the TransplantAssessmentState
we create a deep copy of the RevisionStack object and name it the
landing_stack, where blocked revisions are removed from this stack.
Then all landable paths in the landing_stack from the root revisions
to the leaf revisions are considered to be landable subgraphs. A new
leaf_revisions method is added which returns all nodes that have
no successors in the graph. Each node in the original stack object has a
blocked attribute which is initialized to an empty list. When a
blocking check fails on a revision the reason is appended to the
list in the full revision stack.