checks: split `assess_stack_state` into a state builder and check runner (Bug 1896501)
The `assess_stack_state` function is doing two things at the same time.
Firstly, it is creating the `StackAssessmentState` from some arguments.
Then, it runs the checks against the state and returns a `StackAssessment`.
Split the `assess_stack_state` function into a function that returns a
`StackAssessmentState`, and have callers call `run_landing_checks`
directly. This will enable us to move the `create_state` function into
a proper fixture in a later PR.