hg: add a snippet to consider Treestatus API failures as closed (Bug 1896696)
Our current snippet parsing for `hg push` output looks for `is CLOSED`
to determine if a tree is closed. When the Treestatus API returns an
error, it will instead print that is it "treating as if CLOSED". Since
we don't check for that specific string we can sometimes treat closures
as permanent failures. Add a snippet to account for this tree-closed
state.