checks: only block assessment when revision is in landing path (Bug 1897007)
After deploying the checks refactor, stacks with closed parents are
being shown as unable to land. In the old implementation of checks,
a subset of the checks applied to every revision in the stack, but
did not apply to the returned assessment. In the new implementation,
we are currently only running checks on the "relevant" revisions,
which means all revisions for regular stack checks and only the
landing path revisions for landing assessment checks. The latter
logic has the side-effect of not checking revisions outside of the
landing path, including the closed parents, which results in the
"valid landing path" check failing since it appears the passed
landing path is trying to land part of the stack without including
an open parent.

To resolve this, we first add two tests that reproduce the bug
conditions showing rejected landings with open and closed parents.

We modify and rename the valid paths check to assert the passed
landing path is a simple path within the graph before running any
checks using the `nx.is_simple_path` algorithm. We add a check that
the landing path is a valid path and special case it to run after
all other checks have completed. With this logic we can identify
paths that are invalid within the graph because they are not linked,
and paths that are invalid because they are not landable due to the
blocked state of the graph.

Next we split `revision_check_pairs` into a function which returns
every revision in the stack and a function which returns the revisions
which are relevant for assessment blocks. We change the logic within
the revision-wise blockers loop to avoid adding the blocker to the
assessment when the revision being checked is not in the relevant set,
but continue removing the revision from the landable stack.

After these changes, the `latest_diff` check may not include all
revisions in the `revision_to_diff_id` field, since it is built up
from the landing path, so we make the retrieval of the diff ID
use `.get` and consider a missing revision a non-fatal error.