hooks: exclude treestatus API endpoints from maintenance mode (Bug 1896696) (#407)
Add an exception for the Treestatus API endpoints to the
maintenance mode hook. Unlike the dockerflow endpoints, the
Treestatus endpoints are defined with Connexion, and the
derived value of `request.endpoint` ends up being an underscore
separated represenation of the function namespace. Add a check
to exclude any endpoints that are within that namespace.