hooks: add a diff inspection API for landings and try pushes (Bug 1899584) (#417)
Add a `DiffAssessor` class as an internal API for running
diff-based checks. Move existing diff-inspecting checks
into the API and update landing checks to use them. Add
a `run_diff_checks` method which runs each check and
returns a list of string errors for display on failure.
Hook the `run_diff_checks` API into the try push revision
building function, so the currently ported checks are
run on try pushes. Add integration tests to show that
pushes are accepted/rejected as appropriate based on
these diff checks.

Since some checks will require inspection of the landing
repo to determine if the check is valid, we add a
`get_repo_for_revision` method that returns the repo for
a revision based on the passed `revision` Phabricator API
object. The logic is derived from the uplift blocking
check where it is updated, and is re-used in the
`try_task_config.json` blocker test.