hgexports: port NSS/NSRP and WPTSync bot update prevention hooks as Lando checks (Bug 1903456) (#425)
Add a `check_wpt_sync` check to the `DiffAssessor` class which
is a port of the `prevent_wptsync_changes.py` hook in v-c-t.
The hook uses a regex to assert that the WPTSync bot user cannot
make changes to files outside of a specific directory in the main
tree. The bot is allowed to push any changes to `try`, but is
otherwise only allowed to make changes to files matching the regex.

Implement the `prevent_nspr_nss_changes.py` hook from v-c-t as a
check on the `DiffAssessor` API within Lando. The hook checks for
file changes under given directories in the tree and prevents updates
without a specific keyword in the commit message.