requirements: update celery, redis and kombu (Bug 1904121) (#427)
Due to a regression in newer Python build tools, our
currently installed version of `celery` fails to install
in the Docker image build as a result of improperly formatted
version specifiers. Install the latest version of `celery`
to resolve the issue, which also requires an update to
the `kombu` dependency. Updating `kombu` then causes a
regression in `redis`, so update `redis` to the latest
`3.x` version.

In newer versions of `celery` the pytest fixtures are disabled
by default, so add a `pytest_plugins` variable to the top-level
`conftest.py` to enable it.

Celery commands now fail when `sys.argv[0]` is passed as part
of `argv`, so remove them instead. The worker also requires the
`worker` command as the first argument, so update the call to
add it. We also add the broker URL as the `results_backend` in
the worker config.