hg: convert errors from `hg id` to `HgException` (Bug 1908492) (#431)
Since `hg id <repo_url>` hits the network to connect to hgmo,
it can fail due to internal server errors and other connectivity
issues. In this case, we want the landing job to be retried in
the same fashion as when failing to pull/push. Map the exception
raised by `run_hg` to an `HgException` so the failure will be
considered a retry in the landing worker.