git-patch-helper: set charset to `utf-8` in email message parser (Bug 1893976)
When parsing the Git patch sent through `git format-patch`, the
email parser returns escaped unicode. Explicitly set the charset
of the email to `utf-8` so the unescaped unicode string is returned.

Add a tests which uses a real-life example of this issue, which fails
without specifying the charset but passes afterwards.