checks: add a push-wide check API and implement bug references check (Bug 1915671) (#432)
Adds an API for push-wide checks and implements the bug references check
using it. The API defines a `PushCheck` interface which initializes the
check with a repo, defines a `next_diff()` function which accepts each diff 
parsed as a `PatchHelper`, and a `result()` function which returns the error
message for the check, or `None` if the check passed.

The `BugReferencesCheck` is the first check to be implemented using this
new API. The check is only relevant on `try` repositories. Each diff is
passed via `next_diff`, and if `SKIP_BMO_CHECK` is found in the commit
message of any diff, a flag is set which indicates the check will be
skipped, otherwise the bug numbers found in the commit message are
stored. On `result()` the `skip_check` flag will cause an unconditional
pass, otherwise BMO is pinged to check the bug numbers in the commit
message are all referencing public bugs.

Since the `author` field was always assumed to be an email address, it
was renamed to the more accurate `email` name.