bmo: use `common_headers` instead of `headers` in requests (Bug 1925420) (#437)
`headers` includes headers passed to the callers, which we add to the
`common_headers` dict to include things like the UA string and
API key. When we create the actual request, we are instead passing
the `headers` object and ignoring our common headers. Pass
`common_headers` to `requests.request` instead.