patch-helper: account for non-Git formatted author headers (Bug 1932215)
`parse_git_author_information` is using `email.utils.parseaddr` to
parse the authorship information from Git and Mercurial patches.
In Mercurial, the standard Git format is not required, so users can
use values like `ffxbld` as their author header. When the parser
encounters fields such as this, the improperly formatted name
is returned as the email.

Update `parse_git_author_information` to return the raw header
as the username when the email is parsed as empty.