Bug 1944462 - Properly handle patches containing CRLF in GitPatchHelper
HgPatchHelper doesn't have this problem because it uses io.StringIO,
which yields the line endings.

Some tests need to be accommodated to the fact that the last line of a
diff now does contain newlines, which was accidentally not the case
before, and while doing that, adjust a little more test strings to be
more fitting with the reality that a patch will not terminate without
a newline at the end.