try_push: add support for passing Git base SHAs (Bug 1960791)
Add support for passing a Git base SHA in the try push
API endpoint. A new `base_commit_format` field is added
to the request body, where the client can pass `git` or `hg`
as the base commit format.

When a landing job is processed with a `target_commit_hash_vcs`
of `git`, it will use `git cinnabar git2hg` to convert the
Git SHA to the equivalent Mercurial SHA. To facilitate this
lookup, we install `git-cinnabar` in the Docker image and
add basic Git repo handling capabilities to the `HgRepo`
class. The class will add a Git repo from a `native_git_source`
specified on the `Repo` object, and when passed it will clone
a Git repo in a sibling directory to the main Mercurial repo
from the source. After the initial cloning, we will fetch
data using Cinnabar grafting to enable lookups of SHAs with
`git cinnabar git2hg`.